app/requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
psycopg2-binary==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
six==1.15.0
starlette==0.13.6
StringGenerator==0.4.3
typing-extensions==3.7.4.3
uvicorn==0.13.4
